+++ plugins/module_utils/network/iosxr/facts/facts.py
@@ -80,6 +80,9 @@ from ansible_collections.cisco.iosxr.plugins.module_utils.network.iosxr.facts.pr
 from ansible_collections.cisco.iosxr.plugins.module_utils.network.iosxr.facts.logging_global.logging_global import (
     Logging_globalFacts,
 )
+from ansible_collections.cisco.iosxr.plugins.module_utils.network.iosxr.facts.ntp_global.ntp_global import (
+    Ntp_globalFacts,
+)
 
 
 FACT_LEGACY_SUBSETS = dict(
@@ -105,6 +108,7 @@ FACT_RESOURCE_SUBSETS = dict(
     bgp_global=Bgp_globalFacts,
     prefix_lists=Prefix_listsFacts,
     logging_global=Logging_globalFacts,
+    ntp_global=Ntp_globalFacts,
 )